Fix jest mock for python_package
