+++ pyscriptjs/package.json
@@ -2,13 +2,13 @@
     "name": "pyscript",
     "version": "0.0.1",
     "scripts": {
-        "build": "npm run tsc && node esbuild.js",
-        "dev": "NODE_WATCH=1 node esbuild.js",
+        "build": "npm run tsc && node esbuild.mjs",
+        "dev": "NODE_WATCH=1 node esbuild.mjs",
         "tsc": "tsc --noEmit",
-        "format:check": "prettier --check './src/**/*.{js,html,ts}'",
-        "format": "prettier --write './src/**/*.{js,html,ts}'",
-        "lint": "eslint './src/**/*.{js,html,ts}'",
-        "lint:fix": "eslint --fix './src/**/*.{js,html,ts}'",
+        "format:check": "prettier --check './src/**/*.{mjs,js,html,ts}'",
+        "format": "prettier --write './src/**/*.{mjs,js,html,ts}'",
+        "lint": "eslint './src/**/*.{mjs,js,html,ts}'",
+        "lint:fix": "eslint --fix './src/**/*.{mjs,js,html,ts}'",
         "xprelint": "npm run format",
         "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --coverage",
         "test:watch": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --watch"
